Combined bugfixes.
Add missing CancelDomain and UncancelDomain roles.

Allow non-domain arguments to --domain for AccessControlListAdd for EPP client certificates.

Fix bug in check_arguments() that prevented specifying only a --handle-id for DomainCreate and DomainUpdate.

Allow wildcards in HandleId for searching in HandleDetailsQry.

Specify a User-Agent when making HTTP requests.

Bump version to 0.3.